chore(Makefile): Add static build target for grepcidr
This change introduces a new target in the Makefile to create a static
binary of the grepcidr application. This allows users to compile
a version of grepcidr that does not depend on shared libraries,
which can be useful for deployment in environments where shared
libraries are not available.

(cherry picked from commit 0ec6283ebc7b7fa814a3a4b7bd155f268472ecb6)